Benchmark extension baseline performance
